coonect/requirements/base.txt
Django==1.10.6
django-cors-headers==2.0.0
djangorestframework==3.6.2
django-extensions==1.7.5
django-filter==1.0.1
django-mysql==1.1.0
mysqlclient==1.3.9
Pillow==4.0.0
pytz==2016.10
opentok==2.4.1
pydocs==0.2
pubnub==4.0.9
celery==4.0.2
braintree==3.35.0
